thêm đếm bước cho 4 thuật toán sort, quick,merge,radix, heap
- đếm bước cho thêm 4 sort,cần phảimở db set compare =1 cho 4 thuật toán sort còn lại trc khi run
-trang newsdetail có thể xem được với tư cách guest

+++ AVS/src/java/Model/NewsModel.java
@@ -107,9 +107,9 @@ public class NewsModel {
     }
 
     //delete cua like comment by comment id
-    public void deleteLikeNewsByNewsId(int newsid) throws Exception {
+    public void deleteLikeNewsByNewsId(int newsid,int userid) throws Exception {
         //excute update
-        String query = "DELETE FROM LikeNews WHERE news_ID = ?";
+        String query = "DELETE FROM LikeNews WHERE news_ID = ? and user_ID= ?";
         DBContext dbManager = new DBContext();
         Connection conn = null;
         PreparedStatement ps = null; //de nhan paramenter
@@ -117,6 +117,7 @@ public class NewsModel {
             conn = dbManager.getConnection();
             ps = conn.prepareStatement(query);
             ps.setInt(1, newsid);
+            ps.setInt(2, userid);
             int executeUpdate;
             executeUpdate = ps.executeUpdate();
         } catch (SQLException e) {
